pembaharuan kegiatan sub 1 s/d 4 dan ref angka kredit

+++ src/models/refKegiatanSub1.js
@@ -32,7 +32,7 @@ const RefKegiatanSub1 = db.define(
             allowNull: false
         },
         angka_kredit: {
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.FLOAT(11),
             allowNull: true
         },
         ucr: {
